refactor: 퀴즈 상태 초기화 동작 useQuizPhaseClear hook으로 분리 및 반영
