+++ src/components/quiz/molecules/QuizPlayButtonGroup/index.tsx
@@ -4,7 +4,7 @@ import { memo } from 'react';
 
 import ButtonGroup from '@/components/@common/molecules/ButtonGroup';
 import Modal from '@/components/@common/molecules/Modal';
-import useQuizResult from '@/hooks/quiz/useQuizResult';
+import useQuizPhaseClear from '@/hooks/quiz/useQuizPhaseClear';
 import useModal from '@/hooks/useModal';
 import useRouteChangeBlocking from '@/hooks/useRouteChangeBlocking';
 
@@ -16,7 +16,7 @@ interface QuizPlayButtonGroupProps {
 const QuizPlayButtonGroup = memo(({ succeedText, onSucceed }: QuizPlayButtonGroupProps) => {
   const { visible: isConfirmModalVisible, open: openConfirmModal, close: closeConfirmModal } = useModal();
   const router = useRouter();
-  const { clearQuizPhase } = useQuizResult();
+  const clearQuizResult = useQuizPhaseClear();
   const { offRouteChangeBlocking } = useRouteChangeBlocking(openConfirmModal);
   return (
     <>
@@ -24,7 +24,7 @@ const QuizPlayButtonGroup = memo(({ succeedText, onSucceed }: QuizPlayButtonGrou
       {isConfirmModalVisible && (
         <Modal visible={isConfirmModalVisible} onClose={closeConfirmModal} hasCloseIcon={false}>
           <Modal.Body>현재 페이지로 다시 돌아올 수 없습니다?</Modal.Body>
-          <Modal.ConfirmButton onClick={() => offRouteChangeBlocking(clearQuizPhase)}>나가기</Modal.ConfirmButton>
+          <Modal.ConfirmButton onClick={() => offRouteChangeBlocking(clearQuizResult)}>나가기</Modal.ConfirmButton>
           <Modal.CloseButton onClick={closeConfirmModal}>취소</Modal.CloseButton>
         </Modal>
       )}
